Fixed imports + decomposed canvas
