+++ src/shared/components/BannerCanvas/index.ts
@@ -1,0 +1,3 @@
+import BannerCanvas from './ui/BannerCanvas'
+
+export { BannerCanvas }
